+++ ezomero/__init__.py
@@ -1,13 +1,4 @@
-from ._ezomero import (get_image,
-                       get_image_ids,
-                       get_map_annotation_ids,
-                       get_map_annotation,
-                       get_file_annotation_ids,
-                       get_file_annotation,
-                       get_group_id,
-                       get_user_id,
-                       get_original_filepaths,
-                       put_map_annotation,
+from ._ezomero import (put_map_annotation,
                        connect,
                        store_connection_params,
                        set_group)
@@ -24,6 +15,15 @@ from ._posts import (post_dataset,
                      post_file_annotation,
                      post_project,
                      post_roi)
+from ._gets import (get_image,
+                    get_image_ids,
+                    get_map_annotation_ids,
+                    get_map_annotation,
+                    get_file_annotation_ids,
+                    get_file_annotation,
+                    get_group_id,
+                    get_user_id,
+                    get_original_filepaths)
 
 __all__ = ['post_dataset',
            'post_image',
